Extract ShardRule so all screenshot tests are actually sharded
The shard filter was an anonymous TestRule declared only in
WearScreenshotTest. Because it is a JUnit rule rather than a runner or a
Gradle-level filter, it only applied to that class's subclasses. The other
screenshot base classes are siblings, not subclasses, so their tests ran
in every shard:

  WearScreenshotTest                             6 classes  sharded
  WearLegacyScreenTest (: WearScreenshotTest)   37 classes  sharded
  WearLegacyA11yTest                            36 classes  NOT sharded
  WearLegacyComponentTest                       45 classes  NOT sharded

That is 81 of 124 classes running six times over, so the six-way matrix
did 81*6 + 43 = 529 class-runs instead of 124 and bought almost no
wall-clock. It also meant a single failing screenshot was reported by
every shard at once.

@Category(ScreenshotTest::class) had the identical bug - it was likewise
declared only on WearScreenshotTest, so the build job's
-Dtest.excludeCategories did not exclude those 81 classes either.

Extracts the rule into ShardRule and declares it, along with the category,
on WearLegacyA11yTest and WearLegacyComponentTest.
WearTriadScreenshotTest and WearDeviceScreenshotTest already extend
WearScreenshotTest and are unaffected.

Verified on :composables across all six shards: every screenshot class is
now assigned to exactly one shard, and DatePickerA11yTest runs only in
shard 5 instead of all six.

Note that assignment is still className.hashCode() mod totalShards, which
balances poorly. Fixing that needs the full class list, which a TestRule
cannot see, so it would mean moving sharding into a Gradle-level filter -
left as a follow-up, which would delete this rule rather than improve it.

TAG=agy
CONV=709d690e-2237-4eaa-bef0-ab88bbe85909
